Extract traversal methods from tree/mod.rs into tree/traversal.rs
Move FlatNode struct and internal traversal helpers (flatten,
flatten_node, flatten_node_filtered, subtree_matches, get_node,
get_node_mut, expand_all_recursive, collapse_all_recursive,
revalidate_selection) to a dedicated submodule. This reduces
mod.rs from 997 to 817 lines, keeping it well under the
1000-line project limit.
